change definition of cardsLIst
