+++ frontend/src/components/EditView/CardsList/CardListProps.types.ts
@@ -1,0 +1,6 @@
+import type { CardProps } from "../Card/CardProps.types";
+
+export interface CardListProps {
+    cards: CardProps[],
+    deleteCard: (cardId: number) => void
+}
